chore(react): pulled out the folder
