+++ packages/react/src/index.ts
@@ -1,1 +1,5 @@
-export * from "./lib";
+export { FibrForm } from "./FibrForm";
+export { FibrProvider } from "./FibrProvider";
+export { Fields } from "./Fields";
+export { RenderField } from "./RenderField";
+export type { FieldProps } from "./types";
