--- packages/react/src/lib/index.ts
@@ -1,5 +1,0 @@
-export { FibrForm } from "./FibrForm";
-export { FibrProvider } from "./FibrProvider";
-export { Fields } from "./Fields";
-export { RenderField } from "./RenderField";
-export type { FieldProps } from "./types";
